requirements.txt
cachetools==5.5.2
deprecation==2.1.0
grpcio==1.70.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
protobuf==4.25.6
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b109
tzdata==2025.1


tinkoff-investments>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiofiles>=22.0.0  # Для асинхронной работы с файлами